Fix 41 medium and low bugs: pricing, performance, UX, mobile, WCAG, code quality
MEDIUM (BUG-031 to BUG-054):
- BUG-031: parseFloat for decimal prices in special pricing
- BUG-032: Enum whitelist for review status updates
- BUG-033: Promise.all for parallel dashboard queries
- BUG-034: Pagination limit on RevenueDetail queries
- BUG-035/046: UTC-based night calculations (DST-safe)
- BUG-036: Clear coupon on date change
- BUG-037: Default currency to SEK instead of USD
- BUG-038: Minimum guest validation in payment function
- BUG-039: Normalized recommended sort (rating 70% + value 30%)
- BUG-040: Shared MAX_PRICE_RANGE constant
- BUG-041: Typed MapBounds with validation
- BUG-042: Fallback route when API key missing
- BUG-043: Toast listener subscribes once on mount
- BUG-044: DOMPurify sanitization on guestName
- BUG-045: Error code check in auth service
- BUG-047: Validate checkout URL before redirect
- BUG-048: Email change verification note
- BUG-049: 60s cooldown on password reset
- BUG-050: Booking ownership check in chat
- BUG-051: Separate bookings loading state
- BUG-052: React Router navigate in PropertyGuide
- BUG-053: Differentiated z-index for mobile nav layers
- BUG-054: Sticky submit button on mobile booking form

LOW (BUG-055 to BUG-071):
- BUG-055/056: Typo fixes (Nordic, platform)
- BUG-057: Crypto-secure CSRF nonce
- BUG-058: Error state for failed payments
- BUG-059: AlertDialog in YaleLockManagement
- BUG-060: LRU cache eviction (max 100 entries)
- BUG-061: Throw on missing Supabase config
- BUG-062: WCAG 44px touch targets
- BUG-063: Gallery dot contrast + ARIA labels
- BUG-064: Responsive hero text sizing
- BUG-065: Single month calendar on mobile
- BUG-066: Dynamic viewport height for dialogs
- BUG-067: SEO page titles on 5 pages
- BUG-068: Remove dead imports
- BUG-069: Remove unnecessary email sanitization
- BUG-070: Proper type casts replacing 'as any'
- BUG-071: Correct amenity count display

+++ src/components/admin/PropertySpecialPricingEnhanced.tsx
@@ -75,7 +75,7 @@ export const PropertySpecialPricingEnhanced = ({
 
     setLoading(true);
     try {
-      const price = parseInt(specialPrice) || null;
+      const price = parseFloat(specialPrice) || null;
 
       for (const date of selectedDates) {
         const { error } = await supabase.from("availability").upsert(
